requirements.txt
# Core dependencies
python-dotenv>=1.0.0
neo4j>=5.14.0
PyPDF2>=3.0.0
plantuml>=0.3.0
pdfplumber>=0.10.0

# NLP and ML dependencies
spacy>=3.6.0
scikit-learn>=1.3.0
nltk>=3.8.1

# LLM and API dependencies
openai>=1.0.0
requests>=2.31.0

# Utility dependencies
typing-extensions>=4.7.0
python-json-logger>=2.0.7
tqdm>=4.66.1

# Validation and evaluation dependencies
networkx>=3.1.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
difflib>=1.0.0
matplotlib>=3.7.0

# Additional dependencies for the new functionality
python-dateutil>=2.8.2 